Fixed bug when selecting a new renderer from the menu bar(and lots more)
* Fixed a bug that happened when selecting a renderer from "Window ->
Applications"; the new renderer would be set as active, but then SWT
would fire another widgetSelected event for the MenuItem that was
previously selected, so the renderer that was selected before becomes
the active renderer *again*, causing the user to be unable to change
renderers using the menu (but the keyboard shortcuts still worked)
* Added and changed the location of "@since 1.0" in all classes
* Added more descriptions to some of the JavaDocs in the Mouse class and
tweaked the deltaXY logic a bit
+ Added more descriptions to some of the JavaDocs in the Keyboard class
+ Had the GLThread print OpenGL version and renderer information on
startup (mostly for debugging purposes right now)
* Made the UICallback class' methods check to see that the window is not
closed before continuing (otherwise, with multiple windows open, if one
window is closed, its UICallback will still attempt to listen to events
and can throw exceptions)
* Changed how the VideoHelper class creates FloatBuffers
+ Added and implemented MenuProvider.updateMenuItems()
+ Added and implemented Renderer.onCleanup()
+ Started implementing ControllerManager
+ Added and implemented InputCallback.onControllerConnected(Controller)
and onControllerDisconnected(Controller) (needs testing!)
+ Added some more utility functions to the GLUtil class
* Tweaked the way the GLUtil.matrix4x4ToString(...) methods format the
whitespace for a nicer-looking printout
* Modified the default "missing icon" image provided by the
SWTResourceManager class. Note that if you use the SWT WindowBuilder,
this class may be overwritten and reset back to its default state. To
prevent this, just right click the class in Eclipse and mark it as
'Read-Only' under the 'Resource' section once it has been updated.
+ More updates to follow; this commit is getting rather large.