Created a new class named "MatrixStack"
Created "MatrixStack" to make it easier to manage matrices when
modifying scenes/objects/cameras in OpenGL.

This class' primary purpose is to add support for forward-compatible
contexts which lack the deprecated GL functions, such as
GL11.glLoadMatrix(...), GL11.glPushMatrix(), etc. by allowing you to
manipulate a matrix and then pass it on to your shader(s) instead.